% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obesity.R
\docType{data}
\name{obesity}
\alias{obesity}
\title{Incomplete clustered data related to obesity}
\format{
\subsection{\code{obesity}}{

A data frame with 2,111 rows and 7 columns:
\describe{
\item{cluster}{Grouping variable}
\item{gender}{Gender (0 = male, 1 = female)}
\item{age}{Age (in years)}
\item{height}{Height (in meters)}
\item{favc}{Frequent consumption of high caloric food (0 = no, 1 = yes)}
\item{weight}{Weight (in kilograms)}
\item{time}{Response time (in minutes)}
}
}
}
\source{
\url{https://www.kaggle.com/datasets/fabinmndez/obesitydata}
}
\usage{
obesity
}
\description{
Data adapted from https://www.kaggle.com/datasets/fabinmndez/obesitydata.
}
\keyword{datasets}
